chore(1-3336): extract selectable periods
